requirements.txt
Pillow==9.1.1
icrawler==0.6.6
opencv-python==4.6.0.66
numpy==1.23.0
moviepy==1.0.3
